+++ crates/block-producer/src/block_builder/mod.rs
@@ -1,20 +1,22 @@
 use std::{collections::BTreeSet, ops::Range};
 
-use miden_node_utils::formatting::format_array;
+use futures::FutureExt;
 use miden_objects::{
     account::AccountId,
     batch::ProvenBatch,
-    block::Block,
-    note::{NoteHeader, Nullifier},
+    block::{Block, BlockNumber},
+    note::{NoteHeader, NoteId, Nullifier},
     transaction::{InputNoteCommitment, OutputNote},
 };
+use opentelemetry::trace::Status;
 use rand::Rng;
 use tokio::time::Duration;
-use tracing::{debug, info, instrument};
+use tracing::{instrument, Span};
+use tracing_opentelemetry::OpenTelemetrySpanExt;
 
 use crate::{
-    errors::BuildBlockError, mempool::SharedMempool, store::StoreClient, COMPONENT,
-    SERVER_BLOCK_FREQUENCY,
+    block::BlockInputs, errors::BuildBlockError, mempool::SharedMempool, store::StoreClient,
+    COMPONENT, SERVER_BLOCK_FREQUENCY,
 };
 
 pub(crate) mod prover;
@@ -32,7 +34,7 @@ pub struct BlockBuilder {
     /// Simulated block failure rate as a percentage.
     ///
     /// Note: this _must_ be sign positive and less than 1.0.
-    pub failure_rate: f32,
+    pub failure_rate: f64,
 
     pub store: StoreClient,
     pub block_kernel: BlockProver,
@@ -72,37 +74,153 @@ impl BlockBuilder {
         loop {
             interval.tick().await;
 
-            let (block_number, batches) = mempool.lock().await.select_block();
-
-            let mut result = self.build_block(&batches).await;
-            let proving_duration = rand::thread_rng().gen_range(self.simulated_proof_time.clone());
-
-            tokio::time::sleep(proving_duration).await;
-
-            // Randomly inject failures at the given rate.
-            //
-            // Note: Rng::gen rolls between [0, 1.0) for f32, so this works as expected.
-            if rand::thread_rng().gen::<f32>() < self.failure_rate {
-                result = Err(BuildBlockError::InjectedFailure);
-            }
-
-            let mut mempool = mempool.lock().await;
-            match result {
-                Ok(_) => mempool.block_committed(block_number),
-                Err(_) => mempool.block_failed(block_number),
-            }
+            self.build_block(&mempool).await;
         }
     }
 
-    #[instrument(target = COMPONENT, skip_all, err)]
-    async fn build_block(&self, batches: &[ProvenBatch]) -> Result<(), BuildBlockError> {
-        info!(
-            target: COMPONENT,
-            num_batches = batches.len(),
-            batches = %format_array(batches.iter().map(ProvenBatch::id)),
-        );
+    #[instrument(parent = None, target = COMPONENT, name = "block_builder.build_block", skip_all)]
+    async fn build_block(&self, mempool: &SharedMempool) {
+        use futures::TryFutureExt;
 
-        let updated_account_set: BTreeSet<AccountId> = batches
+        // Run the block building stages and add open-telemetry trace information where applicable.
+        //
+        // Important telemetry is added to the root span, and stages are free to add additional data
+        // to their subspans. Each stage has its own span and is expected to emit an error
+        // event if it fails. This automatically marks the span as an error, however the
+        // root span must be manually marked as this function does not return a result.
+        Self::select_block(mempool)
+            .inspect(SelectedBlock::inject_telemetry)
+            .then(|selected| self.get_block_inputs(selected))
+            .inspect_ok(BlockPreimage::inject_telemetry)
+            .and_then(|inputs| self.prove_block(inputs))
+            .inspect_ok(ProvenBlock::inject_telemetry)
+            // Failure must be injected before the final pipeline stage i.e. before commit is called. The system cannot
+            // handle errors after it considers the process complete (which makes sense).
+            .and_then(|proven_block| async { self.inject_failure(proven_block) })
+            .and_then(|proven_block| self.commit_block(mempool, proven_block))
+            // Handle errors by propagating the error to the root span and rolling back the block.
+            .inspect_err(|err| Span::current().set_status(Status::Error { description: format!("{err:?}").into() }))
+            .or_else(|_err| self.rollback_block(mempool).never_error())
+            // Error has been handled, this is just type manipulation to remove the result wrapper.
+            .unwrap_or_else(|_| ())
+            .await;
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.select_block", skip_all)]
+    async fn select_block(mempool: &SharedMempool) -> SelectedBlock {
+        let (block_number, batches) = mempool.lock().await.select_block();
+        SelectedBlock { block_number, batches }
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.get_block_inputs", skip_all, err)]
+    async fn get_block_inputs(
+        &self,
+        selected_block: SelectedBlock,
+    ) -> Result<BlockPreimage, BuildBlockError> {
+        let SelectedBlock { block_number: _, batches } = selected_block;
+        let summary = BlockSummary::summarize_batches(&batches);
+
+        let inputs = self
+            .store
+            .get_block_inputs(
+                summary.updated_accounts.iter().copied(),
+                summary.nullifiers.iter(),
+                summary.dangling_notes.iter(),
+            )
+            .await
+            .map_err(BuildBlockError::GetBlockInputsFailed)?;
+
+        let missing_notes: Vec<_> = summary
+            .dangling_notes
+            .difference(&inputs.found_unauthenticated_notes.note_ids())
+            .copied()
+            .collect();
+        if !missing_notes.is_empty() {
+            return Err(BuildBlockError::UnauthenticatedNotesNotFound(missing_notes));
+        }
+
+        Ok(BlockPreimage { batches, summary, inputs })
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.prove_block", skip_all, err)]
+    async fn prove_block(&self, preimage: BlockPreimage) -> Result<ProvenBlock, BuildBlockError> {
+        let BlockPreimage { batches, summary, inputs } = preimage;
+
+        let (block_header_witness, updated_accounts) = BlockWitness::new(inputs, &batches)?;
+
+        let new_block_header = self.block_kernel.prove(block_header_witness)?;
+
+        let block = Block::new(
+            new_block_header,
+            updated_accounts,
+            summary.output_notes,
+            summary.nullifiers,
+        )
+        .expect("invalid block components");
+
+        self.simulate_proving().await;
+
+        Ok(ProvenBlock { block })
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.commit_block", skip_all, err)]
+    async fn commit_block(
+        &self,
+        mempool: &SharedMempool,
+        proven_block: ProvenBlock,
+    ) -> Result<(), BuildBlockError> {
+        self.store
+            .apply_block(&proven_block.block)
+            .await
+            .map_err(BuildBlockError::StoreApplyBlockFailed)?;
+
+        mempool.lock().await.commit_block();
+
+        Ok(())
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.rollback_block", skip_all)]
+    async fn rollback_block(&self, mempool: &SharedMempool) {
+        mempool.lock().await.rollback_block();
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.simulate_proving", skip_all)]
+    async fn simulate_proving(&self) {
+        let proving_duration = rand::thread_rng().gen_range(self.simulated_proof_time.clone());
+
+        Span::current().set_attribute("range.min_s", self.simulated_proof_time.start.as_secs_f64());
+        Span::current().set_attribute("range.max_s", self.simulated_proof_time.end.as_secs_f64());
+        Span::current().set_attribute("dice_roll_s", proving_duration.as_secs_f64());
+
+        tokio::time::sleep(proving_duration).await;
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.inject_failure", skip_all, err)]
+    fn inject_failure<T>(&self, value: T) -> Result<T, BuildBlockError> {
+        let roll = rand::thread_rng().gen::<f64>();
+
+        Span::current().set_attribute("failure_rate", self.failure_rate);
+        Span::current().set_attribute("dice_roll", roll);
+
+        if roll < self.failure_rate {
+            Err(BuildBlockError::InjectedFailure)
+        } else {
+            Ok(value)
+        }
+    }
+}
+
+struct BlockSummary {
+    updated_accounts: BTreeSet<AccountId>,
+    nullifiers: Vec<Nullifier>,
+    output_notes: Vec<Vec<OutputNote>>,
+    dangling_notes: BTreeSet<NoteId>,
+}
+
+impl BlockSummary {
+    #[instrument(target = COMPONENT, name = "block_builder.summarize_batches", skip_all)]
+    fn summarize_batches(batches: &[ProvenBatch]) -> Self {
+        let updated_accounts: BTreeSet<AccountId> = batches
             .iter()
             .flat_map(ProvenBatch::account_updates)
             .map(|(account_id, _)| *account_id)
@@ -111,7 +229,7 @@ impl BlockBuilder {
         let output_notes: Vec<_> =
             batches.iter().map(|batch| batch.output_notes().to_vec()).collect();
 
-        let produced_nullifiers: Vec<Nullifier> =
+        let nullifiers: Vec<Nullifier> =
             batches.iter().flat_map(ProvenBatch::produced_nullifiers).collect();
 
         // Populate set of output notes from all batches
@@ -120,8 +238,8 @@ impl BlockBuilder {
             .flat_map(|output_notes| output_notes.iter().map(OutputNote::id))
             .collect();
 
-        // Build a set of unauthenticated input notes for this block which do not have a matching
-        // output note produced in this block
+        // Build a set of unauthenticated input notes for this block which do not have a
+        // matching output note produced in this block
         let dangling_notes: BTreeSet<_> = batches
             .iter()
             .flat_map(ProvenBatch::input_notes)
@@ -130,47 +248,79 @@ impl BlockBuilder {
             .filter(|note_id| !output_notes_set.contains(note_id))
             .collect();
 
-        // Request information needed for block building from the store
-        let block_inputs = self
-            .store
-            .get_block_inputs(
-                updated_account_set.into_iter(),
-                produced_nullifiers.iter(),
-                dangling_notes.iter(),
-            )
-            .await
-            .map_err(BuildBlockError::GetBlockInputsFailed)?;
-
-        let missing_notes: Vec<_> = dangling_notes
-            .difference(&block_inputs.found_unauthenticated_notes.note_ids())
-            .copied()
-            .collect();
-        if !missing_notes.is_empty() {
-            return Err(BuildBlockError::UnauthenticatedNotesNotFound(missing_notes));
+        Self {
+            updated_accounts,
+            nullifiers,
+            output_notes,
+            dangling_notes,
         }
+    }
+}
 
-        let (block_header_witness, updated_accounts) = BlockWitness::new(block_inputs, batches)?;
+struct SelectedBlock {
+    block_number: BlockNumber,
+    batches: Vec<ProvenBatch>,
+}
+struct BlockPreimage {
+    batches: Vec<ProvenBatch>,
+    summary: BlockSummary,
+    inputs: BlockInputs,
+}
+struct ProvenBlock {
+    block: Block,
+}
 
-        let new_block_header = self.block_kernel.prove(block_header_witness)?;
+impl SelectedBlock {
+    fn inject_telemetry(&self) {
+        let span = Span::current();
+        span.set_attribute("block.number", i64::from(self.block_number.as_u32()));
+        span.set_attribute("block.batches.count", i64::from(self.batches.len() as u32));
+    }
+}
 
-        // TODO: return an error?
-        let block =
-            Block::new(new_block_header, updated_accounts, output_notes, produced_nullifiers)
-                .expect("invalid block components");
+impl BlockPreimage {
+    fn inject_telemetry(&self) {
+        let span = Span::current();
 
-        let block_hash = block.hash();
-        let block_num = new_block_header.block_num();
+        // SAFETY: We do not expect to have more than u32::MAX of any count per block.
+        span.set_attribute(
+            "block.updated_accounts.count",
+            i64::try_from(self.summary.updated_accounts.len())
+                .expect("less than u32::MAX account updates"),
+        );
+        span.set_attribute(
+            "block.output_notes.count",
+            i64::try_from(self.summary.output_notes.len())
+                .expect("less than u32::MAX output notes"),
+        );
+        span.set_attribute(
+            "block.nullifiers.count",
+            i64::try_from(self.summary.nullifiers.len()).expect("less than u32::MAX nullifiers"),
+        );
+        span.set_attribute(
+            "block.dangling_notes.count",
+            i64::try_from(self.summary.dangling_notes.len())
+                .expect("less than u32::MAX dangling notes"),
+        );
+    }
+}
 
-        info!(target: COMPONENT, %block_num, %block_hash, "block built");
-        debug!(target: COMPONENT, ?block);
+impl ProvenBlock {
+    fn inject_telemetry(&self) {
+        let span = Span::current();
+        let header = self.block.header();
 
-        self.store
-            .apply_block(&block)
-            .await
-            .map_err(BuildBlockError::StoreApplyBlockFailed)?;
+        span.set_attribute("block.hash", header.hash().to_hex());
+        span.set_attribute("block.sub_hash", header.sub_hash().to_hex());
+        span.set_attribute("block.parent_hash", header.prev_hash().to_hex());
 
-        info!(target: COMPONENT, %block_num, %block_hash, "block committed");
+        span.set_attribute("block.protocol.version", i64::from(header.version()));
 
-        Ok(())
+        span.set_attribute("block.roots.kernel", header.kernel_root().to_hex());
+        span.set_attribute("block.roots.nullifier", header.nullifier_root().to_hex());
+        span.set_attribute("block.roots.account", header.account_root().to_hex());
+        span.set_attribute("block.roots.chain", header.chain_root().to_hex());
+        span.set_attribute("block.roots.note", header.note_root().to_hex());
+        span.set_attribute("block.roots.transation", header.tx_hash().to_hex());
     }
 }