+++ crates/block-producer/src/mempool/mod.rs
@@ -9,7 +9,7 @@ use miden_objects::{
     transaction::TransactionId,
     MAX_ACCOUNTS_PER_BATCH, MAX_INPUT_NOTES_PER_BATCH, MAX_OUTPUT_NOTES_PER_BATCH,
 };
-use tokio::sync::Mutex;
+use tokio::sync::{Mutex, MutexGuard};
 use tracing::instrument;
 use transaction_expiration::TransactionExpirations;
 use transaction_graph::TransactionGraph;
@@ -127,7 +127,15 @@ impl BlockBudget {
 // MEMPOOL
 // ================================================================================================
 
-pub type SharedMempool = Arc<Mutex<Mempool>>;
+#[derive(Clone)]
+pub struct SharedMempool(Arc<Mutex<Mempool>>);
+
+impl SharedMempool {
+    #[instrument(target = COMPONENT, name = "mempool.lock", skip_all)]
+    pub async fn lock(&self) -> MutexGuard<'_, Mempool> {
+        self.0.lock().await
+    }
+}
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct Mempool {
@@ -168,13 +176,13 @@ impl Mempool {
         state_retention: usize,
         expiration_slack: u32,
     ) -> SharedMempool {
-        Arc::new(Mutex::new(Self::new(
+        SharedMempool(Arc::new(Mutex::new(Self::new(
             chain_tip,
             batch_budget,
             block_budget,
             state_retention,
             expiration_slack,
-        )))
+        ))))
     }
 
     fn new(
@@ -205,7 +213,7 @@ impl Mempool {
     /// # Errors
     ///
     /// Returns an error if the transaction's initial conditions don't match the current state.
-    #[instrument(target = COMPONENT, skip_all, fields(tx=%transaction.id()))]
+    #[instrument(target = COMPONENT, name = "mempool.transactions.add", skip_all, fields(tx=%transaction.id()))]
     pub fn add_transaction(
         &mut self,
         transaction: AuthenticatedTransaction,
@@ -227,7 +235,7 @@ impl Mempool {
     /// Transactions are returned in a valid execution ordering.
     ///
     /// Returns `None` if no transactions are available.
-    #[instrument(target = COMPONENT, skip_all)]
+    #[instrument(target = COMPONENT, name = "mempool.batch.select", skip_all)]
     pub fn select_batch(&mut self) -> Option<(BatchId, Vec<AuthenticatedTransaction>)> {
         let (batch, parents) = self.transactions.select_batch(self.batch_budget);
         if batch.is_empty() {
@@ -243,7 +251,7 @@ impl Mempool {
     /// Drops the failed batch and all of its descendants.
     ///
     /// Transactions are placed back in the queue.
-    #[instrument(target = COMPONENT, skip_all, fields(batch))]
+    #[instrument(target = COMPONENT, name = "mempool.batch.failed", skip_all, fields(batch_id=%batch))]
     pub fn batch_failed(&mut self, batch: BatchId) {
         // Batch may already have been removed as part of a parent batches failure.
         if !self.batches.contains(&batch) {
@@ -267,7 +275,7 @@ impl Mempool {
     }
 
     /// Marks a batch as proven if it exists.
-    #[instrument(target = COMPONENT, skip_all, fields(batch=%batch.id()))]
+    #[instrument(target = COMPONENT, name = "mempool.batch.proved", skip_all, fields(batch_id=%batch.id()))]
     pub fn batch_proved(&mut self, batch: ProvenBatch) {
         // Batch may have been removed as part of a parent batches failure.
         if !self.batches.contains(&batch.id()) {
@@ -286,7 +294,7 @@ impl Mempool {
     /// # Panics
     ///
     /// Panics if there is already a block in flight.
-    #[instrument(target = COMPONENT, skip_all)]
+    #[instrument(target = COMPONENT, name = "mempool.block.select", skip_all)]
     pub fn select_block(&mut self) -> (BlockNumber, Vec<ProvenBatch>) {
         assert!(self.block_in_progress.is_none(), "Cannot have two blocks inflight.");
 
@@ -300,11 +308,9 @@ impl Mempool {
     ///
     /// # Panics
     ///
-    /// Panics if blocks are completed out-of-order or if there is no block in flight.
-    #[instrument(target = COMPONENT, skip_all, fields(block_number))]
-    pub fn block_committed(&mut self, block_number: BlockNumber) {
-        assert_eq!(block_number, self.chain_tip.child(), "Blocks must be submitted sequentially");
-
+    /// Panics if there is no block in flight.
+    #[instrument(target = COMPONENT, name = "mempool.block.commit", skip_all)]
+    pub fn commit_block(&mut self) {
         // Remove committed batches and transactions from graphs.
         let batches = self.block_in_progress.take().expect("No block in progress to commit");
         let transactions =
@@ -321,7 +327,7 @@ impl Mempool {
         self.chain_tip = self.chain_tip.child();
 
         // Revert expired transactions and their descendents.
-        let expired = self.expirations.get(block_number);
+        let expired = self.expirations.get(self.chain_tip);
         self.revert_transactions(expired.into_iter().collect())
             .expect("expired transactions must be part of the mempool");
     }
@@ -330,12 +336,9 @@ impl Mempool {
     ///
     /// # Panics
     ///
-    /// Panics if there is no block in flight or if the block number does not match the current
-    /// inflight block.
-    #[instrument(target = COMPONENT, skip_all, fields(block_number))]
-    pub fn block_failed(&mut self, block_number: BlockNumber) {
-        assert_eq!(block_number, self.chain_tip.child(), "Blocks must be submitted sequentially");
-
+    /// Panics if there is no block in flight.
+    #[instrument(target = COMPONENT, name = "mempool.block.rollback", skip_all)]
+    pub fn rollback_block(&mut self) {
         let batches = self.block_in_progress.take().expect("No block in progress to be failed");
 
         // Revert all transactions. This is the nuclear (but simplest) solution.
@@ -370,6 +373,7 @@ impl Mempool {
     ///
     /// Returns an error if any transaction was not in the transaction graph i.e. if the transaction
     /// is unknown.
+    #[instrument(target = COMPONENT, name = "mempool.transactions.revert", skip_all)]
     fn revert_transactions(
         &mut self,
         txs: Vec<TransactionId>,